+++ src/categories/js/stylelint/config/default.config.ts
@@ -3,20 +3,7 @@ import { prettierMutation } from '../stylelint.utils';
 import { Config } from './config.interface';
 
 const stylelintConfig: Config['stylelintConfig'] = {
-  extends: ['stylelint-config-standard', 'stylelint-config-rational-order-fix'],
-  plugins: ['stylelint-order', 'stylelint-config-rational-order-fix/plugin'],
-  rules: {
-    'declaration-empty-line-before': null,
-    'no-empty-first-line': null,
-    'order/properties-order': [],
-    'plugin/rational-order': [
-      true,
-      {
-        'border-in-box-model': false,
-        'empty-line-between-groups': true,
-      },
-    ],
-  },
+  extends: ['stylelint-config-standard', 'stylelint-config-clean-order'],
 };
 
 const stylelintIgnore = readableMultilineString`
@@ -28,7 +15,7 @@ const stylelintIgnore = readableMultilineString`
 export const defaultConfig: Config = {
   stylelintConfig,
   stylelintIgnore,
-  devDependencies: ['stylelint', 'stylelint-config-standard', 'stylelint-order', 'stylelint-config-rational-order-fix'],
+  devDependencies: ['stylelint', 'stylelint-config-standard', 'stylelint-config-clean-order'],
   scripts: [
     { name: 'csslint', script: 'stylelint "src/**/*.css"' },
     { name: 'csslint:fix', script: 'stylelint "src/**/*.css" --fix' },